+++ settee/ViewController.swift
@@ -95,17 +95,29 @@ class ViewController: UITableViewController, CLLocationManagerDelegate, TWTRTwee
     }
 
     func loadTweets(tweetIDs: [String]) {
+        if tweetIDs.isEmpty {
+            return
+        }
+
         // Do not trigger another request if one is already in progress.
         if self.isLoadingTweets {
             return
         }
 
+        self.isLoadingTweets = true
+
         // load tweets with guest login
         Twitter.sharedInstance().logInGuestWithCompletion { (session: TWTRGuestSession!, error: NSError!) in
+            if session == nil {
+                self.isLoadingTweets = false
+                println("Twitter guest login failed")
+                return
+            }
 
             // Find the tweets with the tweetIDs
             Twitter.sharedInstance().APIClient.loadTweetsWithIDs(tweetIDs) {
                 (twttrs, error) -> Void in
+                self.isLoadingTweets = false
 
                 // If there are tweets do something magical
                 if ((twttrs) != nil) {